+++ src/components/Auth/AuthInput/AuthInputBox/AuthPwInputBox.tsx
@@ -1,5 +1,5 @@
-import AuthInputWrapper from './AuthInputWrapper';
-import AuthInput from './AuthInput';
+import AuthInputWrapper from '../AuthInputWrapper';
+import AuthInput from '../AuthInput';
 import { UseFormRegister, UseFormResetField } from 'react-hook-form';
 
 interface Props {
