+++ src/components/Auth/AuthInput/AuthInputBox/index.ts
@@ -1,0 +1,5 @@
+import AuthEmailInputBox from './AuthEmailInputBox';
+import AuthPwInputBox from './AuthPwInputBox';
+import AuthPwCheckInputBox from './AuthPwCheckInputBox';
+
+export { AuthEmailInputBox, AuthPwInputBox, AuthPwCheckInputBox };
