+++ src/components/Auth/index.ts
@@ -1,13 +1,13 @@
-import AuthInput from './AuthInput/AuthInput';
-import AuthEmailInputBox from './AuthInput/AuthEmailInputBox';
-import AuthPwInputBox from './AuthInput/AuthPwInputBox';
-import AuthPwCheckInputBox from './AuthInput/AuthPwCheckInputBox';
+import {
+  AuthEmailInputBox,
+  AuthPwInputBox,
+  AuthPwCheckInputBox,
+} from './AuthInput/AuthInputBox';
 import ErrorMessage from './ErrorMessage';
 import KakaoLoginButton from './KakaoLoginButton';
 import SignupButton from './SignupButton';
 
 export {
-  AuthInput,
   AuthEmailInputBox,
   AuthPwInputBox,
   AuthPwCheckInputBox,
